+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed ignoreBins config option
+
 ## v0.4.8 (2023-12-06)
 * Added ignoreBins option to config